Replace calls to os.get_terminal_size with shutil.get_terminal_size
On Linux, os.get_terminal_size only works if Blender is started from a
terminal. Trying to render a spritesheet when not running from a
terminal will cause an exception. Using shutil.get_terminal_size fixes
this issue since it returns a fallback size (80 x 24) when not running
in a terminal.